Cardio train project
In many data-driven applications, information is stored in CSV (Comma-Separated Values) files, which may contain unorganized or unsorted data. This project focuses on developing a tool that reads data from a CSV file, sorts it based on specified criteria, and outputs a well-organized file. This project is essential for ensuring data consistency, facilitating easier data analysis, and improving data readability